Bclerc

+++ back/src/wschat/wschat.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { WebSocketServer } from '@nestjs/websockets';
-import { ChatRoom, Message, PenaltyType, User } from '@prisma/client';
+import { ChatPenalty, ChatRoom, Message, PenaltyType, User } from '@prisma/client';
 import { ChatService } from 'src/chat/chat.service';
 import { SubscribeRoomDto } from 'src/chat/dto/subscribe-room.dto';
-import { ChatRoomI, MessageI, newChatRoomI } from 'src/chat/interfaces/chatRoom.interface';
+import { ChatRoomI, MessageI, newChatRoomI, PardonI } from 'src/chat/interfaces/chatRoom.interface';
 import { EjectRoomI } from 'src/chat/interfaces/eject-room-i.interface';
 import { DemoteUserI, PromoteUserI } from 'src/chat/interfaces/promote-user-i.interface';
 import { PasswordUtils } from 'src/chat/utils/chat-utils';
@@ -14,6 +14,7 @@ import { EventEmitter2 } from '@nestjs/event-emitter';
 import { PenaltiesService } from 'src/chat/services/penalties/penalties.service';
 import e from 'express';
 import { PusnishI } from 'src/chat/interfaces/punish.interface';
+import { BlockedUser } from 'src/chat/interfaces/blocked.interface';
 
 @Injectable()
 export class WschatService {
@@ -41,35 +42,35 @@ export class WschatService {
   async subscribeToRoom(socketId: string, subRoom: SubscribeRoomDto) {
     const user = this.onlineUserService.getUser(socketId);
     const room = await this.chatService.getRoomById(subRoom.roomId);
-    const penalty = await this.penaltiesService.getRoomPenaltiesForUser(user.id, room.id);    
-    console.log(penalty);
-    if (penalty && penalty.type === PenaltyType.BAN) {
-      this.eventEmitter.emit('room.user.join', {
-        room: room,
-        user: user,
-        success: false,
-        message: 'Vous avez été banni de ce salon' + (penalty.timetype === 'PERM' ? '.' : ' jusqu\'au ' + penalty.endTime.toLocaleString('fr-FR'))
-      });
-      return ;
-    }
-
-    if (user) {
-      if (!await this.chatService.canJoin(room.id, subRoom.password)) {
+    const penalty = await this.penaltiesService.getRoomPenaltiesForUser(user.id, room.id);   
+    
+    if (user && room) {
+      if (penalty && penalty.type === PenaltyType.BAN) {
         this.eventEmitter.emit('room.user.join', {
           room: room,
           user: user,
           success: false,
-          message: 'Mot de passe incorrect',
-         });
-         return ;
+          message: 'Vous avez été banni de ce salon' + (penalty.timetype === 'PERM' ? '.' : ' jusqu\'au ' + penalty.endTime.toLocaleString('fr-FR'))
+        });
+        return ;
       }
 
-      await this.chatService.addUsersToRoom(room.id, user.id);
-      this.eventEmitter.emit('room.user.join', {
-        room: room,
-        user: user,
-        success: true,
-      });
+        if (!await this.chatService.canJoin(room.id, subRoom.password)) {
+          this.eventEmitter.emit('room.user.join', {
+            room: room,
+            user: user,
+            success: false,
+            message: 'Mot de passe incorrect',
+          });
+          return ;
+        }
+
+        await this.chatService.addUsersToRoom(room.id, user.id);
+        this.eventEmitter.emit('room.user.join', {
+          room: room,
+          user: user,
+          success: true,
+        });
     }
   }
 
@@ -81,7 +82,7 @@ export class WschatService {
     let endTime: Date;
 
     if (user && target && room) {
-      if (room.admins.find(admin => admin.id == user.id)) {
+      if (room.admins.find(admin => admin.id == user.id) || user.id == room.ownerId) {
         if (room.ownerId != target.id) {
           endTime = new Date(now + event.time);
           this.penaltiesService.punishUser(target.id, room.id, event.type, event.perm ? null : endTime);
@@ -98,13 +99,27 @@ export class WschatService {
     }
   }
 
+
+  async pardonUser(socketId: string, event: PardonI) {
+    const user = this.onlineUserService.getUser(socketId);
+    const target = await this.userService.findOne(event.userId);
+    const penalty: any = await this.penaltiesService.getPenaltyById(event.penaltyId);
+    
+    if (user && target && penalty) {
+          const room = penalty.room;
+          if (room.admins.find(admin => admin.id == user.id) || user.id == room.ownerId) {
+            this.penaltiesService.deletePenalty(penalty.id);
+            this.eventEmitter.emit('room.user.pardoned', { room: room, user: target, pardoner: user, success: true});
+          }       
+      }
+  }
+
   async ejectUserFromRoom(socketId: string, room: EjectRoomI) {
     const user = this.onlineUserService.getUser(socketId);
     const target = await this.userService.findOne(room.targetId);
     const roomToEject = await this.chatService.getRoomById(room.roomId);
-
-    if (user && target.id != roomToEject.ownerId) {
-      if (roomToEject.admins.find(admin => admin.id == user.id)) {
+    if (user && target && roomToEject && target.id != roomToEject.ownerId) {
+      if (roomToEject.admins.find(admin => admin.id == user.id) || user.id == roomToEject.ownerId) {
         await this.chatService.removeUsersFromRoom(roomToEject.id, target.id);
         this.eventEmitter.emit('room.user.kicked', { room: roomToEject, user: target, kicker: user });
       }
@@ -116,7 +131,7 @@ export class WschatService {
     const target = await this.userService.findOne(event.targetId);
     const roomToPromote = await this.chatService.getRoomById(event.roomId);
 
-    if (user) {
+    if (user && target && roomToPromote) {
       if (user.id == roomToPromote.ownerId) {
         this.chatService.addAdminsToRoom(roomToPromote.id, target.id);
         this.updateRoomForUsersInRoom(roomToPromote.id);
@@ -135,7 +150,7 @@ export class WschatService {
     const target = await this.userService.findOne(event.targetId);
     const roomToDemote = await this.chatService.getRoomById(event.roomId);
 
-    if (user) {
+    if (user && target && roomToDemote) {
       if (roomToDemote.ownerId == user.id && roomToDemote.ownerId != target.id) {
         this.chatService.removeAdminsFromRoom(roomToDemote.id, target.id);
         this.updateRoomForUsersInRoom(roomToDemote.id);
@@ -154,9 +169,13 @@ export class WschatService {
     const user = this.onlineUserService.getUser(socketId);
     const room = await this.chatService.getRoomById(message.room.id);
 
-    if (user) {
-      await this.chatService.newMessage(message);
-      this.eventEmitter.emit('room.message.new', { room: room });
+    if (user && room) {
+      try {
+        await this.chatService.newMessage(message);
+        this.eventEmitter.emit('room.message.new', { room: room });
+      } catch (error) {
+        this.eventEmitter.emit('room.user.canchat', { room: room, user: user, message: error.response});
+      }
     }
   }
 
@@ -178,10 +197,10 @@ export class WschatService {
       const newRoom = await this.chatService.removeUsersFromRoom(room.id, user.id);
       if (room.ownerId == user.id) {
         if (newRoom.users.length > 0) {
-          this.chatService.updateRoomOwner(room.id, newRoom.users[0].id);
-          this.updateRoomForUsersInRoom(room.id);
+         await this.chatService.updateRoomOwner(room.id, newRoom.users[0].id);
         } else {
           // delete
+
           this.eventEmitter.emit('room.delete', { room: room });
         }
       }
@@ -194,15 +213,15 @@ export class WschatService {
     const room = await this.chatService.getRoomById(newRoom.id);
     if (!newRoom.name
         || newRoom.name.length < 3
-        || newRoom.name.length > 20
+        || newRoom.name.length > 30
         || !newRoom.description
         || newRoom.description.length < 3
-        || newRoom.description.length > 100)
+        || newRoom.description.length > 200)
       {
         this.eventEmitter.emit('room.update', {user: user, room: room, success: false, message: 'Les champs ne sont pas valides' });
         return ;
       }
-    if (user) {
+    if (user && room) {
       if (room.ownerId == user.id) {
         await this.chatService.editRoom(newRoom);
         this.eventEmitter.emit('room.update', {user: user, room: newRoom, success: true });
@@ -212,11 +231,26 @@ export class WschatService {
     }
   }
 
+  async blockUser(socketId: string, event: BlockedUser) {
+    const user = this.onlineUserService.getUser(socketId);
+    const target = await this.userService.findOne(event.userId);
+
+    if (user && target && user.id != target.id) {
+      if (event.block) {
+        await this.userService.blockUser(user.id, target.id);
+        this.eventEmitter.emit("user.blocked", { user: target, blocker: user, block: true, success: true });
+      } else {
+        await this.userService.unblockUser(user.id, target.id);
+        this.eventEmitter.emit("user.blocked", { user: target, blocker: user, block: false, success: true });
+      }
+    }
+  }
+
   async updateRoomForUsersInRoom(roomId: number) {
     let room = await this.chatService.getRoomById(roomId);
 
     for (let user of room.users) {
-      let rooms = await this.chatService.getRoomsFromUser(user.id);
+      const rooms = await this.chatService.getRoomsFromUser(user.id);
       this.sendToUser(user, 'rooms', rooms);
     }
   }
